PromQL: inline the relabel translation
Translating label_replace/label_join through a separate RelabelLayout
class spread one translation over two files for no gain: the layout had
a single caller, and reading the translator meant chasing the derivation
into another class. Bring the derivation back into the translator as the
methods it was before - labelReplaceValue, labelJoinValue,
sourceLabelValue - next to doTranslateMetadataManipulation.

The labels a relabel reads stay a property of
MetadataManipulationFunction (sourceLabels), where the analyzer's shadow
resolution and the translator both need them.
